+++ EOLib/Packet/InitSequenceStart.cs
@@ -48,7 +48,7 @@ public sealed class InitSequenceStart : ISequenceStart
     {
         int value = random.Next(1757);
         int seq1Max = (value + 13) / 7;
-        int seq1Min = Math.Max(0, (value - (EoNumericLimits.CHAR_MAX - 1) + 13 + 6) / 7);
+        int seq1Min = Math.Max(0, (value - ((int)EoNumericLimits.CHAR_MAX - 1) + 13 + 6) / 7);
 
         int seq1 = random.Next(seq1Max - seq1Min) + seq1Min;
         int seq2 = value - seq1 * 7 + 13;
